# gatewayservice/openapi.yaml
openapi: 3.0.0
info:
  title: Gatewayservice API
  description: Gateway OpenAPI specification.
  version: 0.2.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: http://${{ secrets.DEPLOY_HOST }}:8000
    description: Production server
paths:
  /adduser:
    post:
      summary: Add a new user to the database.
      operationId: addUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: User ID.
                  example: student
                password:
                  type: string
                  description: User password.
                  example: pass
                name:
                  type: string
                  description: User name.
                  example: Pepe
                surname:
                  type: string
                  description: User surname.
                  example: Grillo
      responses:
        '200':
          description: User added successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                    description: User ID
                  password:
                    type: string
                    description: Hashed password
                    example: $2b$10$ZKdNYLWFQxzt5Rei/YTc/OsZNi12YiWz30JeUFHNdAt7MyfmkTuvC
                  _id:
                    type: string
                    description: Identification
                    example: 65f756db3fa22d227a4b7c7d
                  createdAt:
                    type: string
                    description: Creation date.
                    example: '2024-03-17T20:47:23.935Z'
                  ___v:
                    type: integer
                    example: '0'                                        
        '400':
          description: Failed to add user.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: getaddrinfo EAI_AGAIN mongodb
  /health:
    get:
      summary: Check the health status of the service.
      operationId: checkHealth
      responses:
        '200':
          description: Service is healthy.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: Health status.
                    example: OK
  /login:
    post:
      summary: Log in to the system.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: User ID.
                  example: student
                password:
                  type: string
                  description: User password.
                  example: pass
      responses:
        '200':
          description: Login successful. Returns user token, username, and creation date.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: User token.
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2NWY3NTZkYjNmYTIyZDIyN2E0YjdjN2QiLCJpYXQiOjE3MTA3MDg3NDUsImV4cCI6MTcxMDcxMjM0NX0.VMG_5DOyQ4GYlJQRcu1I6ICG1IGzuo2Xuei093ONHxw
                  username:
                    type: string
                    description: Username.
                    example: student
                  createdAt:
                    type: string
                    description: Creation date.
                    example: '2024-03-17T20:47:23.935Z'
        '401':
          description: Invalid credentials.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Shows the error info..
                    example: Invalid credentials
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /generate-question:
    get:
      summary: Generate a new question.
      operationId: generateQuestion
      responses:
        '200':
          description: Question generated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  question:
                    type: string
                    description: Question.
                    example: ¿Cuál es la capital de Francia?
                  correctAnswer:
                    type: string
                    description: Answer.
                    example: Paris
                  allAnswers:
                    type: array
                    description: Array with all possible answers.
                    items:
                      type: string
                      example:
                        - Paris
                        - Madrid
                        - Berlin
                        - Roma
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /questions:
    get:
      summary: Get all questions.
      operationId: getQuestions
      responses:
        '200':
          description: Questions retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                      description: Question ID.
                      example: 65f756db3fa22d227a4b7c7d
                    question:
                      type: string
                      description: Question.
                      example: ¿Cuál es la capital de Francia?
                    correctAnswer:
                      type: string
                      description: Answer.
                      example: Paris
                    allAnswers:
                      type: array
                      description: Array with all possible answers.
                      items:
                        type: string
                        example:
                          - Paris
                          - Madrid
                          - Berlin
                          - Roma
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /users:
      get:
        summary: Get all users. 
        operationId: getUsers
        responses:
          '200':
            description: Users retrieved successfully.
            content:
              application/json:
                schema:
                  type: array
                  items:
                    type: object
                    properties:
                      _id:
                        type: string
                        description: user ID.
                        example: 65f756db3fa22d227a4b7c7d
                      username:
                        type: string
                        description: User.
                        example: carmen
                      gamesPlayed:
                        type: integer
                        description: Games played.
                        example: 2
                      rigthAnswers:
                        type: integer
                        description: Right answers.
                        example: 5
                      wrongAnswers:
                        type: integer
                        description: Wrong answers.
                        example: 5
          
          '500':
            description: Internal server error.
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    error:
                      type: string
                      description: Error information.
                      example: Internal Server Error